# docs/swagger.yaml
basePath: /api/
definitions:
...
  types.CreateOrderRequest:
    properties:
      addonTotal:
        type: number
      customerId:
        type: string
      paymentMethod:
        description: e.g. "paymongo", "cash"
        type: string
      quoteId:
        type: string
      subtotal:
        type: number
      totalAmount:
        type: number
    required:
    - addonTotal
    - customerId
    - paymentMethod
    - quoteId
    - subtotal
    - totalAmount
    type: object
  types.CreateOrderResponse:
    properties:
      orderId:
        type: string
      orderNumber:
        type: string
    type: object
...
  types.ErrorResponse:
    properties:
      error:
        type: string
    type: object
...
  types.FetchAllQuotesResponse:
    properties:
      quotes:
        items:
          $ref: '#/definitions/types.Quote'
        type: array
      quotesRequested:
        type: integer
      totalQuotes:
        type: integer
    type: object
...
  types.GetOrdersResponse:
    properties:
      orders:
        items:
          $ref: '#/definitions/types.Order'
        type: array
      ordersRequested:
        type: integer
      totalOrders:
        type: integer
    type: object
  types.GetPaymentsResponse:
    properties:
      payments:
        items:
          $ref: '#/definitions/types.Payment'
        type: array
      paymentsRequested:
        type: integer
      totalPayments:
        type: integer
    type: object
...
  types.Order:
    properties:
      addon_total:
        type: number
      created_at:
        type: string
      currency:
        type: string
      customer_id:
        type: string
      downpayment_required:
        type: number
      id:
        type: string
      order_number:
        type: string
      payment_method:
        type: string
      payment_status:
        type: string
      quote_id:
        type: string
      remaining_balance:
        type: number
      subtotal:
        type: number
      total_amount:
        type: number
      updated_at:
        type: string
    type: object
  types.Payment:
    properties:
      amount:
        type: number
      created_at:
        type: string
      currency:
        type: string
      failed_reason:
        type: string
      id:
        type: string
      order_id:
        type: string
      paid_at:
        type: string
      payment_id:
        type: string
      payment_intent_id:
        type: string
      payment_method_id:
        type: string
      provider:
        description: PAYMONGO | CASH | MANUAL
        type: string
      status:
        type: string
      type:
        description: DOWNPAYMENT | FULLPAYMENT |BALANCE | REFUND
        type: string
      updated_at:
        type: string
    type: object
...
  types.PaymentIntentData:
    properties:
      attributes:
        $ref: '#/definitions/types.PaymentIntentAttributes'
      id:
        type: string
      type:
        type: string
    type: object
...
  types.PaymentIntentResponse:
    properties:
      data:
        $ref: '#/definitions/types.PaymentIntentData'
    type: object
...
  types.Quote:
    properties:
      addonTotal:
        type: number
      addons:
        items:
          $ref: '#/definitions/types.QuoteAddon'
        type: array
      createdAt:
        type: string
      customerId:
        type: string
      id:
        type: string
      isValid:
        type: boolean
      mainService:
        description: the main type of service
        type: string
      mainServiceDetail:
        description: added
        type: object
      mainServiceHours:
        description: added
        type: integer
      subtotal:
        type: number
      totalPrice:
        type: number
      totalServiceHours:
        description: added
        type: integer
      updatedAt:
        type: string
    type: object
  types.QuoteAddon:
    properties:
      addonPrice:
        type: number
      createdAt:
        type: string
      id:
        type: string
      quoteId:
        type: string
      serviceDetail:
        description: serialized ServicesRequest
        type: object
      serviceHours:
        type: integer
      serviceType:
        type: string
    type: object
  types.QuoteRequest:
    properties:
      addons:
        description: same here
        items:
          $ref: '#/definitions/types.AddOnRequest'
        type: array
      customerId:
        type: string
      service:
        allOf:
        - $ref: '#/definitions/types.ServicesRequest'
        description: nested structs usually don't need db tags
    type: object
  types.QuoteResponse:
    properties:
      addonTotal:
        type: number
      addons:
        items:
          $ref: '#/definitions/types.AddOnBreakdown'
        type: array
      mainServiceDetail:
        type: object
      mainServiceHours:
        type: integer
      mainServiceName:
        type: string
      mainServiceTotal:
        type: number
      quoteId:
        type: string
      totalPrice:
        type: number
      totalServiceHours:
        type: integer
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: This is the official API documentation for the Handworks Api.
  title: Handworks API
  version: "1.0"
paths:
...
  /payment/customer:
    get:
      consumes:
      - application/json
      description: Retrieve all payments for a specific customer with pagination
      parameters:
      - description: Customer ID
        in: path
        name: id
        required: true
        type: string
      - default: 0
        description: Page number (starting at 0)
        in: query
        name: page
        type: integer
      - default: 10
        description: Number of payments per page
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/types.GetPaymentsResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/types.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/types.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get payments by customer ID
      tags:
      - Payment
  /payment/customer/quotes:
    get:
      consumes:
      - application/json
      description: Retrieve all quotations associated with a specific customer with
        optional date filtering and pagination.
      parameters:
      - description: Customer ID
        in: query
        name: customerId
        required: true
        type: string
      - description: Start date (YYYY-MM-DD)
        in: query
        name: startDate
        type: string
      - description: End date (YYYY-MM-DD)
        in: query
        name: endDate
        type: string
      - default: 0
        description: Page number (starting at 0)
        in: query
        name: page
        type: integer
      - default: 10
        description: Number of quotes per page
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/types.FetchAllQuotesResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/types.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/types.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get all quotations for a customer
      tags:
      - Payment
  /payment/intent/downpayment/{id}:
    post:
      consumes:
      - application/json
      description: Create a PayMongo payment intent for the order's downpayment amount
      parameters:
      - description: Order ID
        in: query
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/types.PaymentIntentResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/types.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/types.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Create downpayment payment intent
      tags:
      - Payment
  /payment/intent/fullpayment/{id}:
    post:
      consumes:
      - application/json
      description: Create a PayMongo payment intent for the order's full remaining
        balance
      parameters:
      - description: Order ID
        in: query
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/types.PaymentIntentResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/types.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/types.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Create full payment payment intent
      tags:
      - Payment
  /payment/order:
    get:
      consumes:
      - application/json
      description: Retrieve a single order by its ID
      parameters:
      - description: Order ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/types.Order'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/types.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/types.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/types.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get order by ID
      tags:
      - Payment
    post:
      consumes:
      - application/json
      description: Create a new order from an accepted quotation
      parameters:
      - description: Order details
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/types.CreateOrderRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/types.CreateOrderResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/types.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/types.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Create an order
      tags:
      - Payment
  /payment/order/customer/{id}:
    get:
      consumes:
      - application/json
      description: Retrieve all orders for a specific customer with pagination
      parameters:
      - description: Customer ID
        in: path
        name: id
        required: true
        type: string
      - default: 0
        description: Page number (starting at 0)
        in: query
        name: page
        type: integer
      - default: 10
        description: Number of orders per page
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/types.GetOrdersResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/types.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/types.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get orders by customer ID
      tags:
      - Payment
  /payment/order/orders:
    get:
      consumes:
      - application/json
      description: Retrieve all orders with pagination
      parameters:
      - default: 0
        description: Page number (starting at 0)
        in: query
        name: page
        type: integer
      - default: 10
        description: Number of orders per page
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/types.GetOrdersResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/types.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/types.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get all orders
      tags:
      - Payment
  /payment/payments/order:
    get:
      consumes:
      - application/json
      description: Retrieve all payments for a specific order with pagination
      parameters:
      - description: Order ID
        in: path
        name: id
        required: true
        type: string
      - default: 0
        description: Page number (starting at 0)
        in: query
        name: page
        type: integer
      - default: 10
        description: Number of payments per page
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/types.GetPaymentsResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/types.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/types.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get payments by order ID
      tags:
      - Payment
  /payment/quote:
    get:
      consumes:
      - application/json
      description: Retrieves a quote by ID that belongs to a specific customer
      parameters:
      - description: Quote ID
        in: query
        name: quoteId
        required: true
        type: string
      - description: Customer ID
        in: query
        name: customerId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/types.Quote'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/types.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/types.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/types.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get quote by ID for a specific customer
      tags:
      - Payment
    post:
      consumes:
      - application/json
      description: Generate a new quotation for a customer
      parameters:
      - description: Quote details
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/types.QuoteRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/types.QuoteResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/types.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/types.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Create a quotation
      tags:
      - Payment
  /payment/quote/preview:
    post:
      consumes:
      - application/json
      description: Generate a new quotation
      parameters:
      - description: Quote details
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/types.QuoteRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/types.QuoteResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/types.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/types.ErrorResponse'
      summary: Create a quotation
      tags:
      - Payment
  /payment/quote/quotes:
    get:
      consumes:
      - application/json
      description: Retrieve all quotations with optional date filtering and pagination.
      parameters:
      - description: Start date (YYYY-MM-DD)
        in: query
        name: startDate
        type: string
      - description: End date (YYYY-MM-DD)
        in: query
        name: endDate
        type: string
      - default: 0
        description: Page number (starting at 0)
        in: query
        name: page
        type: integer
      - default: 10
        description: Number of quotes per page
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/types.FetchAllQuotesResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/types.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/types.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get all quotations
      tags:
      - Payment
securityDefinitions:
  BearerAuth:
    description: Enter "Bearer <your_token>"
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"